+++ src/util/JsonModel.js
@@ -1,0 +1,136 @@
+const fs = require('fs');
+const Collection = require('./Collection');
+
+module.exports = class JsonModel {
+  constructor(db, name, path, defaults) {
+    this.db = db;
+    this.name = name;
+    this.path = path;
+    this.defaults = defaults;
+    this.data = new Collection();
+    let body = {};
+    if (fs.existsSync(path)) {
+      const content = fs.readFileSync(path, 'utf8');
+      try {
+        body = JSON.parse(content);
+      } catch (e) {
+        body = {};
+      }
+    }
+    Object.keys(body).forEach(key =>
+      this.data.set(key, new Document(body[key]))
+    );
+  }
+
+  save() {
+    this.db.save(this.name);
+  }
+
+  filterKeys(query, value) {
+    if (typeof query === 'string') query = { [query]: value };
+    if (typeof query === 'object') {
+      const q = query;
+      query = doc => Object.keys(q).every(k => doc[k] === q[k]);
+    }
+    const keys = [];
+    for (const [key, value] of this.data) {
+      if (query(value, key, this)) keys.push(key);
+    }
+    return keys;
+  }
+
+  filter(query, value) {
+    const keys = this.filterKeys(query, value);
+    const documents = [];
+    for (const key of keys) documents.push(this.data.get(key));
+    return documents;
+  }
+
+  findKey(query, value) {
+    if (typeof query === 'string') query = { [query]: value };
+    if (typeof query === 'object') {
+      const q = query;
+      query = doc => Object.keys(q).every(k => doc[k] === q[k]);
+    }
+    for (const [key, value] of this.data) {
+      if (query(value, key, this)) return key;
+    }
+    return null;
+  }
+
+  findOne(query, value) {
+    const key = this.findKey(query, value);
+    return key ? this.data.get(key) : null;
+  }
+
+  getOne(query, value) {
+    if (typeof query === 'string') {
+      query = { [query]: value };
+    }
+    const defaults = {
+      ...this.defaults,
+      ...(typeof query === 'object' ? query : {}),
+    };
+    return this.findOne(query, value) || new Document(defaults);
+  }
+
+  insertOne(data) {
+    const document = new Document(data);
+    this.data.set(document._id, document);
+    this.save();
+    return document;
+  }
+
+  insertMany(data) {
+    const documents = [];
+    for (const document of data) documents.push(new Document(document));
+    for (const document of documents) this.data.set(document._id, document);
+    this.save();
+    return documents;
+  }
+
+  deleteOne(query, value) {
+    const key = this.findKey(query, value);
+    if (key) {
+      const document = this.data.get(key);
+      this.data.delete(key);
+      this.save();
+      return document;
+    }
+    return null;
+  }
+
+  deleteMany(query, value) {
+    const key = this.findKey(query, value);
+    if (key) {
+      const document = this.data.get(key);
+      this.data.delete(key);
+      this.save();
+      return document;
+    }
+    return null;
+  }
+
+  updateOne(query, value, newData = {}) {
+    const key = this.findKey(query, value);
+    if (!key) return null;
+    if (typeof query !== 'string') newData = value;
+    const document = this.data.get(key);
+    const newDocument = new Document({ ...document, ...newData });
+    this.data.set(key, newDocument);
+    this.save();
+    return newDocument;
+  }
+
+  upsertOne(query, value, newData = {}) {
+    const key = this.findKey(query, value);
+    if (typeof query === 'string') query = { [query]: value };
+    if (!key) {
+      return this.insertOne({
+        ...(typeof query === 'object' ? query : {}),
+        ...newData,
+      });
+    }
+    return this.updateOne(query, value, newData);
+  }
+};